# asm/lexer: Use binary search for keyword lexing