benchmark                old ns/op   new ns/op   delta
BenchmarkParseString     85488       69866       -18.27%