@mewmew 2015-04-22
Beautiful! I had forgotten that the `sort` package provided binary search functions. I am handing in the project this week (finishing up the dissertation right now), so will take a closer look on Friday :)