@quarnster 2015-04-22
Good luck! :)

One addition I thought of afterwards was to just have it as `var keywords []string` and then in the `init` function fill it in by walking over `token.Keywords`. That way you only have one list to maintain and the other is automatically updated.