@mewmew 2015-04-22
> One addition I thought of afterwards was to just have it as var keywords []string and then in the init function fill it in by walking over token.Keywords. That way you only have one list to maintain and the other is automatically updated.

That is a really good idea! I was initially apprehensive to duplicate the defined keywords from the `token` package, as it would duplicate the work of updating the keywords later.

Oh, and thanks! I've mentioned you in the report :) Hope that is ok.