@quarnster 2015-04-23
You're welcome, and I do look forward to continuing to follow this project and perhaps help out here and there. All our communication has been public here so I see no harm in being mentioned in your report :)