@mewmew 2015-04-23
Thanks! I have really appreciated our discussions so far, and however you wish to be involved with this project in the future, it makes me happy to have more people on the team :) My aim is continue developing this library over the next year or so until it reaches production quality. To magically create time for development, I have applied for a couple of Uni units related to compiler development.

>  All our communication has been public here so I see no harm in being mentioned in your report :)

I was using the same reasoning when including your name, whereas private conversations have been anonymized.