@mewmew 2015-04-25
Hi Fredrik,

Now I've handed in the dissertation, and it feels just amazing! I want to merge this pull request and apply the changes you suggested (e.g. populate `keywords` by iterating over `token.Keywords` during `init`).

It seems that I cannot merge this pull request through GitHub (*We can’t automatically merge this pull request.*). I think this may have been my fault as I messed up a commit and had to re-push it. Therefore is seems like the hash of the optimize commit differs from your fork and this repo (llir/llvm@c517e9ef6e85476d910605bd67624d6f29ee45a6 and quarnster/llvm@dfb9d5129ffa2314acf6ea93fae9a5c617d5fd02).

Do you think you could sync with the latest version of the repo, so I can merge the pull request? Feel free to include the changes you mentioned (i.e. populate `keywords` from `token.Keywords`) if you feel like it. Otherwise I will submit that change in a follow-up commit.