@mewmew 2015-04-25
P.S. just in case you are interested and considering your name was mentioned in the report, here is a link to the PDF:
https://raw.githubusercontent.com/mewpaper/decompilation/master/decompilation.pdf