@quarnster 2015-04-27
Did a rebase and had to do a forced push to make anything happen, but think it should be mergeable now. Didn't get a chance to look at the token.Keywords addition nor your report yet.