@mewmew 2015-04-30
Now I've submitted the change which populates keywords from token.Keywords during init. Thanks for the suggestion!