@mewmew 2016-04-21
Hej Fredrik,

How are you? I hope you are well, as I haven't heard from you since last autumn.

Me and a friend of mine are currently implementing a compiler for a subset of C. It's been a happy adventure and we will soon start looking at generating LLVM IR from the semantically verified parse tree.

I would like to ask you if we may reuse the code you wrote to optimize the LLVM IR lexer, as our lexer for C runs into the [exact same performance bottleneck](https://github.com/mewmew/uc/blob/bb746a5a1e01b63c047fc0c3c9855446d20dff1d/hand/lexer/state.go#L317).

Both of these projects [mewmew/uc](https://github.com/mewmew/uc) and [llir/llvm](https://github.com/llir/llvm) have been released into the public domain. We would therefore like to ask if you may consider to release your contribution into the public domain as well.

Hope to hear back from you.

Glada hälsningar -Robin