@mewmew 2017-05-06
@quarnster For public reference, would you be alright with releasing your contribution into the public domain? The intention is to keep the code of the [llir project released into the public domain](https://github.com/llir/llvm#public-domain), to simplify license handling for end users.

Haven't heard from you in a while? How is life nowadays? Hope you have been enjoying the warmth of spring in Sweden :)

Cheers /u